requirements.txt
Django==6.0
django_rq==3.2.1
djangorestframework==3.16.1
redis==7.1.0
Requests==2.32.5
rq==2.6.1
